linux-raspberrypi: Update kernel to 3.12.21
Amongst other patches, this release includes a fix to an issue with gstreamer1.0 and v4l2src blocking

ref: https://bugzilla.gnome.org/show_bug.cgi?id=726521

ref: https://github.com/raspberrypi/linux/commit/fd1355b22e5c52b5e31b936b5f68d3b62cb189c8

To make use of the optional fix the module parameter gst_v4l2src_is_broken needs to be set when loading bcm2835-v4l2.ko

With this in place, and with userland and gstreamer1.0-omx patches, gstreamer1.0 can be used with PiCam via v4l2src instead of raspivid.

(There may still be performance issues to be addressed for v4l2src vs raspivid pipe & fdsrc).

Change-Id: Ia0ed4e6c8f27df9bb12ae2350526f6314e016d51
Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>