weston: Build Weston compositor with rpi-backend
meta-rpi layer now uses userland by default to provide EGL and libGLESv2
libraries needed by Weston. However we are missing a weston-egl library
which would usually be built within Mesa (which we are not using).

As such we disable EGL, build and default to the RPI optimised backend.

This seems to be an optimal solution from a performance perspective but
in future we may wish to add EGL support. X11 support needs testing.

For details on the RPI backend see http://tinyurl.com/pdgkgxo

This patch-set requires the presence of .pc package files in sysroots,
which are provided by a preceding patch-set,

see: I7f6181e0d19ca96968b089f9cdec01223d4b7b58

Change-Id: Ia4a2bcb245a3d21ce133e828a35273a377b07c49
Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>